requirements.txt
feedparser==6.0.11
httpx==0.27.0
deepl==1.18.0
python-telegram-bot==21.5
APScheduler==3.10.4
aiosqlite==0.20.0
rapidfuzz==3.9.7
pytest==8.3.2
pytest-asyncio==0.23.8
tenacity>=8.2
